refactor: 메인페이지 분과용 상수를 전역에서 사용하도록 수정
